% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_webpull.R
\name{eia_webpull}
\alias{eia_webpull}
\title{eia_webpull}
\usage{
eia_webpull(eia_year, dest)
}
\arguments{
\item{eia_year}{year of data to pull from https://www.eia.gov/electricity}

\item{dest}{character string indicating directory where data should be saved. A
subdirectory named for `eia_year` will be created within the directory 
specified and will contain unzipped EIA data files.}
}
\description{
pull EIA thermoelectric plant data including Form EIA-923 (2008-2020), 
            Form EIA-860 (2000-2020), Form EIA-906/920 (2001-2007), 
            Form EIA-767 (2000-2005), and Form EIA-759 (2000 only)
}
\examples{
#set arguments
dest<-tempdir()
eia_year<-2015

#pull data from web
eia_webpull(eia_year,dest)
}
