% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_8D.R
\name{import_8D}
\alias{import_8D}
\title{import_8D}
\usage{
import_8D(cooling8D)
}
\arguments{
\item{cooling8D}{EIA Cooling Operations table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "cooling8D"}
}
\value{
`cooling8D` formatted EIA Cooling Operations table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Cooling Operations table \cr \cr
}
