% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bocoo_data.R
\name{import_bocoo_data}
\alias{import_bocoo_data}
\title{import_bocoo_data}
\usage{
import_bocoo_data(bocoo)
}
\arguments{
\item{bocoo}{EIA Boiler-Cooling association table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "bocoo"}
}
\value{
`bocoo` formatted Boiler-Cooling association table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Boiler-Cooling association table \cr \cr
}
