% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_bogen_data.R
\name{import_bogen_data}
\alias{import_bogen_data}
\title{import_bogen_data}
\usage{
import_bogen_data(bogen)
}
\arguments{
\item{bogen}{EIA Boiler-Generator association table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "bogen"}
}
\value{
`bogen` formatted Boiler-Generator association table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Boiler-Generator association table \cr \cr
}
