% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_boiler_design_data.R
\name{import_boiler_design_data}
\alias{import_boiler_design_data}
\title{import_boiler_design_data}
\usage{
import_boiler_design_data(boilerDesignData)
}
\arguments{
\item{boilerDesignData}{EIA Boiler Design table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "boilerDesignData"}
}
\value{
`boilerDesignData` formatted Boiler Design table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Boiler Design table \cr \cr
}
