% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_boiler_fuel_data.R
\name{import_boiler_fuel_data}
\alias{import_boiler_fuel_data}
\title{import_boiler_fuel_data}
\usage{
import_boiler_fuel_data(boilerFuelData)
}
\arguments{
\item{boilerFuelData}{EIA Boiler Fuel Data table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "boilerFuelData"}
}
\value{
`boilerFuelData` formatted Boiler Fuel Data table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Boiler Fuel Data table \cr \cr
}
