% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_coolingSys.R
\name{import_coolingSys}
\alias{import_coolingSys}
\title{import_coolingSys}
\usage{
import_coolingSys(cooling)
}
\arguments{
\item{cooling}{EIA Cooling System  table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "cooling"}
}
\value{
`cooling` formatted Cooling System  table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Cooling System table \cr \cr
}
