% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_EIAData.R
\name{import_EIAData}
\alias{import_EIAData}
\title{import_EIAData}
\usage{
import_EIAData(
  path_InputData_Metafile,
  path_EIAInputData,
  outputCSV,
  path_outputCSV,
  useStandardCrosswalk = F,
  prepCondenserDuty = F,
  eia_year = NA
)
}
\arguments{
\item{path_InputData_Metafile}{Character string indicating path to crosswalk file must be in "xlsx" 
format. If `useStandardCrosswalk==TRUE` then this is the path to which 
the crosswalk will be saved for further edit by the user.}

\item{path_EIAInputData}{Character string indicating directory in which EIA data files have been
extracted from zip files.  If using the \code{\link{eia_webpull}} function, the 
directory will be specified by the `eia_year`}

\item{outputCSV}{TRUE/FALSE indicating whether or not to save formatted data as individual csv files}

\item{path_outputCSV}{Path to directory in which to save the formatted EIA data as csv files.}

\item{useStandardCrosswalk}{TRUE/FALSE indicating whether the \code{\link{crosswalk2015}} file should be
used to start the crosswalk development process.  After the first run the 
resulting crosswalk will be saved as `path_InputData_Metafile`.  Subsequent 
runs of import_EIAData should have `useStandardCrosswalk` as FALSE.}

\item{prepCondenserDuty}{TRUE/FALSE indicating whether checks should be made for data required to
execute the `EIAplantAssocation` and/or `condenserDuty` packages}

\item{eia_year}{numeric value indicating the year of EIA data being formatted. ONLY required if 
`prepCondenserDuty` is TRUE}
}
\value{
`inputData.list` list of data.frames with standardized column names for columns found in the 
                        crosswalk.  If `prepCondenserDuty=TRUE` the inputData.list will be a list of 
                        the 11 data.frames shown below:\cr \cr
                       
\item{\code{bogen}}{data.frame of 7454 obs of 7 variables sourced from 2015 Form EIA-860 Data - Schedule 6A, 'Boiler / Generator Associations}
\item{\code{bocoo}}{data.frame of	3007 obs. of  7 variables sourced from 2015 Form EIA-860 Data - Schedule 6A, 'Boiler / Cooling System Associations}
\item{\code{gen_860}}{data.frame 20068 obs. of  72 variables sourced from 2015 Form EIA-860 Data - Schedule 3, 'Generator Data' (Operable Units Only)}
\item{\code{retiredGenerators}}{data.frame of 3629 obs. of  56 variables sourced from 2015 Form EIA-860 Data - Schedule 3, 'Generator Data' (Retired & Canceled Units Only)}
\item{\code{generation.data}}{data.frame of 4177 obs. of  28 variables sourced from 2015 Form EIA-923 Monthly Generating Unit Net Generation Time Series File}
\item{\code{boilerFuelData}}{data.frame	of 10001 obs. of  66 variables sourced from 2015 Form EIA-923 Monthly Boiler Fuel Consumption and Emissions Time Series File}
\item{\code{gen_fuel_data}}{data.frame of	12450 obs. of  98 variables sourced from 2015 Form EIA-923 Monthly Generation and Fuel Consumption Time Series File, 2015 Final Revision}  
\item{\code{boilerDesignData}}{data.frame of 4769 obs. of  33 variables sourced from 2015 Form EIA-860 Data - Schedule 6C, 'Boiler Information - Design Parameters}
\item{\code{cooling}}{data.frame of	1892 obs. of  42 variables sourced from 2015 Form EIA-860 Data - Schedule 6D, 'Cooling System Data}
\item{\code{plantInfo}}{data.frame of	8928 obs. of  36 variables sourced from 2015 EIA-860 Data - Plant Info" "Form EIA-860 Data - Plant Info}
\item{\code{cooling8D}}{data.frame of	17159 obs. of  23 variables sourced from 2015 EIA-923, Power Plant Operations Report}
}
\description{
import and format EIA data using a user generated crosswalk or \code{\link{crosswalk2015}}.
            To format data for EIAplantAssociation and/or condenserDuty packages all 
            "scriptImportFunc", "scriptTableName" and "scriptColnames" present in 
            \code{\link{crosswalk2015}} must be present, and all column names of the 
            \code{\link{crosswalk2015}} table must be present.
}
\details{
If the crosswalk is complete and all tables have been imported and formatted, 
       `import_EIAData()` will return a list object containing all tables imported.  If the crosswalk
       is imcomplete/incompatible with the EIA data stored in `path_EIAInputData` the current crosswalk
       file will be modified highlighting missing data and a message will appear indicating that the
       crosswalk file should be edited.  The `import_EIAData()` should be run repeatedly until a list
       of tables is returned.  It is recommended that the resulting crosswalk be saved for future use.
}
\examples{
#set arguments
dest<-tempdir()
eia_year<-2015
path_InputData_Metafile<-paste0(dest,.Platform$file.sep,"UserControlCrosswalk2015.xlsx")
path_EIAInputData<-paste0(dest,.Platform$file.sep,eia_year)
outputCSV<-FALSE
path_outputCSV<-dest

#use crosswalk2015 object
useStandardCrosswalk<-TRUE
prepCondenserDuty<-TRUE

#pull data from web
eia_webpull(eia_year,dest)


#run import using crosswalk2015
inputData.list<-import_EIAData(path_InputData_Metafile,path_EIAInputData,outputCSV,
                              path_outputCSV,prepCondenserDuty = TRUE,eia_year=eia_year,
                              useStandardCrosswalk=useStandardCrosswalk)
}
