% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_generator_data.R
\name{import_generator_data}
\alias{import_generator_data}
\title{import_generator_data}
\usage{
import_generator_data(gen860)
}
\arguments{
\item{gen860}{EIA-860 Generator Data table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "gen860"}
}
\value{
`gen860` formatted Generator Data  table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA-860 Generator Data table \cr \cr
Executed By: import_EIAData() \cr
}
