% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_plantInfo.R
\name{import_plantInfo}
\alias{import_plantInfo}
\title{import_plantInfo}
\usage{
import_plantInfo(plantInfo)
}
\arguments{
\item{plantInfo}{EIA Plant Information table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "plantInfo"}
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Plant Information table \cr \cr
}
