% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sheet1_gen_and_fuel_data.R
\name{import_sheet1_gen_and_fuel_data}
\alias{import_sheet1_gen_and_fuel_data}
\title{import_sheet1_gen_and_fuel_data}
\usage{
import_sheet1_gen_and_fuel_data(gen_fuel_data)
}
\arguments{
\item{gen_fuel_data}{EIA Plant-Reported.Prime.Mover level Fuel Consumption and Generation Data
table import with standardized column names found in \code{\link{crosswalk2015}} 
where scriptTableName = "gen_fuel_data"}
}
\value{
`gen_fuel_data` formatted Plant-Reported.Prime.Mover level Fuel Consumption and Generation 
                       Data  table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Plant-Reported.Prime.Mover level Fuel Consumption and Generation Data table \cr \cr
}
