% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sheet4_gen_and_fuel_data.R
\name{import_sheet4_gen_and_fuel_data}
\alias{import_sheet4_gen_and_fuel_data}
\title{import_sheet4_gen_and_fuel_data}
\usage{
import_sheet4_gen_and_fuel_data(generation.data)
}
\arguments{
\item{generation.data}{EIA Generation Data table import with standardized column names found in 
\code{\link{crosswalk2015}} where scriptTableName = "generation.data"}
}
\value{
`generation.data` formatted Generation Data table
}
\description{
import and format for use by the EIAplantAssociation and condenserDuty packages the 
            EIA Generation Data table\cr \cr
}
