% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importThermoEIA-package.R
\docType{package}
\name{importThermoEIA-package}
\alias{importThermoEIA-package}
\title{Import and Format Thermoelectric EIA data}
\description{
\code{importThermoEIA} pulls data on thermoelectric plants, including Form-923, Form-860, 
Form-767, Form-906920, and Form-759 from the EIA website https://www.eia.gov/electricity/ 
and formats data in a standard format including stable table and column names regardless of 
year data was reported. The importThermoEIA pacakge includes tools to assist the user in 
developing crosswalks for EIA import. The importThermoEIA package formats data to be used by 
the EIAplantAssociation and condenseryDuty packages, but can be used to format EIA data for
custom use outside of the EIAplantAssociation and condenserDuty packages.
}
\details{
For more details see the help vignette:
\code{vignette(package = "importThermoEIA")}

\tabular{ll}{
Package: \tab importThermoEIA\cr
Type: \tab Package\cr
License: \tab Unlimited for this package, dependencies have more restrictive licensing.\cr
Copyright: \tab This software is in the public domain because it contains materials
that originally came from the United States Geological Survey, an agency of
the United States Department of Interior. For more information, see the
official USGS copyright policy at
https://www.usgs.gov/visual-id/credit_usgs.html#copyright\cr
LazyLoad: \tab yes\cr
Disclaimer: \tab This software is preliminary or provisional and is subject to revision. It is being 
provided to meet the need for timely best science. The software has not received final approval by the 
U.S. Geological Survey (USGS). No warranty, expressed or implied, is made by the USGS or the U.S. 
Government as to the functionality of the software and related material nor shall the fact of release 
constitute any such warranty. The software is provided on the condition that neither the USGS nor the 
U.S. Government shall be held liable for any damages resulting from the authorized or unauthorized use 
of the software.\cr
}

\tabular{ll}{
\code{\link{crosswalk2015}} \tab An example of a crosswalk to be used by the EIAplantAssociation
                                and/or condenserDuty packages to import and format EIA data 
                                (https://www.eia.gov/electricity/) necessary to create 
                                boiler-generator-cooling associations and calculate condenser duty 
                                at thermoelectric plants.\cr \cr

\code{\link{eia_webpull}} \tab pull EIA thermoelectric plant data including Form EIA-923 (2008-2020), 
            Form EIA-860 (2000-2020), Form EIA-906/920 (2001-2007), 
            Form EIA-767 (2000-2005), and Form EIA-759 (2000 only)\cr \cr
\code{\link{import_8D}} \tab import and format for use by the EIAplantAssociation and condenserDuty 
                      packages the EIA Cooling Operations table\cr \cr
\code{\link{import_bocoo_data}} \tab import and format for use by the EIAplantAssociation and 
                                condenserDuty packages the EIA Boiler-Cooling association table\cr \cr

\code{\link{import_bogen_data}} \tab import and format for use by the EIAplantAssociation and 
                                condenserDuty packages the EIA Boiler-Generator association table\cr \cr

\code{\link{import_boiler_design_data}} \tab import and format for use by the EIAplantAssociation 
                                      and condenserDuty packages the EIA Boiler Design table\cr \cr

\code{\link{import_boiler_fuel_data}} \tab import and format for use by the EIAplantAssociation and 
                                    condenserDuty packages the EIA Boiler Fuel Data table\cr \cr
\code{\link{import_coolingSys}} \tab import and format for use by the EIAplantAssociation and 
                                    condenserDuty packages the EIA Cooling System table\cr \cr

\code{\link{import_EIAData}} \tab import and format EIA data using a user generated crosswalk or 
                           \code{\link{crosswalk2015}}. To format data for EIAplantAssociation and/or 
                           condenserDuty packages all "scriptImportFunc","scriptTableName" and 
                           "scriptColnames" present in \code{\link{crosswalk2015}} must be present, 
                           and all column names of the \code{\link{crosswalk2015}} table must be 
                           present.\cr \cr

\code{\link{import_generator_data}} \tab import and format for use by the EIAplantAssociation and 
                                  condenserDuty packages the EIA-860 Generator Data table\cr \cr

\code{\link{import_plantInfo}} \tab import and format for use by the EIAplantAssociation and 
                             condenserDuty packages the EIA Plant Information table\cr \cr
                             
\code{\link{import_sheet1_gen_and_fuel_data}} \tab import and format for use by the 
                                                  EIAplantAssociation and condenserDuty packages 
                                                  the EIA Plant-Reported.Prime.Mover level Fuel 
                                                  Consumption and Generation Data table\cr \cr

\code{\link{import_sheet4_gen_and_fuel_data}} \tab import and format for use by the 
                                                  EIAplantAssociation and condenserDuty packages 
                                                  the EIA Generation Data table\cr \cr
                                                  
\code{\link{removeMissing}} \tab replace missing data with 0 and flag row as containing 
                                missing data \cr \cr                                                                                                                                                                   
                                                  
}
}
\keyword{Thermoelectric}
