% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMissing.R
\name{removeMissing}
\alias{removeMissing}
\title{removeMissing}
\usage{
removeMissing(df, cols, missingFlag)
}
\arguments{
\item{df}{data.frame in which to flag missing data}

\item{cols}{numeric vector of column indices to check for missing data or 
character vector of column names}

\item{missingFlag}{character string indicating missing data}
}
\value{
`df` data.frame with missing values flagged and replaced with zeros
}
\description{
replace missing data with 0 and flag row as containing 
            missing data \cr \cr
}
\examples{
df<-data.frame(x = c(1,2,3),y=c(".",100,200))
removeMissing(df,cols=names(df),missingFlag = ".")
}
